Overhaul of boardstate
removed Tiles and 2D array from boardstate, now only contains 2 lists that hold the current pieces. this means that now the maximum searchable area is 32 pieces, which will decrease as pieces are removed throughout the game. this is an improvement from having to search all 64 tiles any time you look for a piece and should significantly improve performance and readability