Update for boardstate changes
updated to work with the new boardstate changes. added convenience function getTileColor to generate the checkerboard pattern without relying on the old Tile class. removed the 10px spacers from the sides of the board. 
TODO - continue updating methods that still rely on Tile so they work with the new boardstate implementation